Improving comments on several modules
