+++ communication/protocols/base/CommunicationProtocolBase.h
@@ -43,11 +43,11 @@ class CommunicationProtocolBase : public cSimpleModule
     // Redirects message to the proper function
     virtual void handleMessage(cMessage *msg);
 
-    // Handles package recieved from communication
+    // Handles package received from communication
     // This packet is a message that was sent to the drone
     virtual void handlePacket(Packet *pk) {};
 
-    // Handles telemetry recieved from mobility
+    // Handles telemetry received from mobility
     // The mobility module exchanges mobility information in the form of telemetry
     virtual void handleTelemetry(Telemetry *telemetry) {};
